requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-multipart
jinja2
